important updates to strcuture better
