+++ mcp_template/backends/kubernetes.py
@@ -1,0 +1,41 @@
+"""
+Kubernetes deployment backend for managing deployments on Kubernetes clusters.
+"""
+
+from typing import Any, Dict, List
+
+from mcp_template.backends import DeploymentBackend
+
+
+class KubernetesDeploymentService(DeploymentBackend):
+    """Kubernetes deployment service (placeholder for future implementation).
+
+    This service will manage Kubernetes deployments when implemented.
+    Currently raises ImportError to indicate it's not yet available.
+    """
+
+    def __init__(self):
+        """Initialize Kubernetes service."""
+        raise ImportError("Kubernetes backend not yet implemented")
+
+    def deploy_template(
+        self,
+        template_id: str,
+        config: Dict[str, Any],
+        template_data: Dict[str, Any],
+        pull_image: bool = True,
+    ) -> Dict[str, Any]:
+        """Deploy template to Kubernetes."""
+        raise NotImplementedError
+
+    def list_deployments(self) -> List[Dict[str, Any]]:
+        """List Kubernetes deployments."""
+        raise NotImplementedError
+
+    def delete_deployment(self, deployment_name: str) -> bool:
+        """Delete Kubernetes deployment."""
+        raise NotImplementedError
+
+    def get_deployment_status(self, deployment_name: str) -> Dict[str, Any]:
+        """Get Kubernetes deployment status."""
+        raise NotImplementedError
