+++ mcp_template/utils/__init__.py
@@ -1,0 +1,8 @@
+"""
+MCP Template Utilities
+"""
+
+from pathlib import Path
+
+ROOT_DIR = Path(__file__).parent.parent.parent
+TEMPLATES_DIR = ROOT_DIR / "templates"
