+++ pytest.ini
@@ -16,7 +16,7 @@ addopts =
     --cov-report=term-missing
     --cov-report=html:htmlcov
     --cov-report=xml:coverage.xml
-    --cov-fail-under=90
+    --cov-fail-under=50
 
 markers =
     unit: Unit tests
