+++ scripts/build_docs.py
@@ -19,7 +19,7 @@ import yaml
 
 # Import the TemplateDiscovery utility
 sys.path.insert(0, str(Path(__file__).parent.parent))
-from mcp_template import TemplateDiscovery
+from mcp_template.template.discovery import TemplateDiscovery
 
 
 def cleanup_old_docs(docs_dir: Path):
