+++ templates/demo/docs/index.md
@@ -1,0 +1,10 @@
+# MCP Demo Server Template
+This template provides a simple demo server for the MCP Platform, showcasing basic functionality and configuration options.
+
+## Overview
+This demo server is designed to demonstrate the capabilities of the MCP Platform architecture, including:
+- FastMCP integration for tool decorators and HTTP transport
+- HTTP-first approach with stdio fallback
+- Docker deployment with networking support
+- Modular code structure following best practices
+- Comprehensive configuration with environment variable mapping
