+++ tests/test_config_strategy.py
@@ -7,9 +7,6 @@ and can be used to test different configuration scenarios.
 """
 
 import json
-import os
-import tempfile
-from pathlib import Path
 
 
 # Mock the config mapper functionality for testing
